# docs/swagger.yaml
basePath: /
definitions:
...
  model.CreatePackingListItemRequest:
    properties:
      box_no_end:
        minimum: 0
        type: integer
      box_no_start:
        minimum: 0
        type: integer
      color:
        type: string
      note:
        type: string
      qty_box:
        type: integer
      qty_per_box:
        type: integer
      sizes:
        items:
          $ref: '#/definitions/model.CreatePackingListItemSizeRequest'
        minItems: 1
        type: array
    required:
    - box_no_end
    - box_no_start
    - color
    - qty_box
    - qty_per_box
    - sizes
    type: object
  model.CreatePackingListItemSizeRequest:
    properties:
      qty:
        type: integer
    required:
    - qty
    type: object
  model.CreatePackingListRequest:
    properties:
      id_surat_jalan_internal:
        type: integer
      id_wo:
        type: integer
      items:
        items:
          $ref: '#/definitions/model.CreatePackingListItemRequest'
        minItems: 1
        type: array
      total_garment_per_box:
        minimum: 0
        type: integer
      total_reject:
        minimum: 0
        type: integer
    required:
    - id_wo
    - items
    - total_garment_per_box
    - total_reject
    type: object
...
  model.CreateSuratJalanClientRequest:
    properties:
      id_material_list:
        type: integer
      keterangan:
        type: string
      qty:
        type: integer
      tanggal:
        type: string
    required:
    - id_material_list
    - qty
    - tanggal
    type: object
...
  model.IssueInventoryRequest:
    properties:
      id_rekonsiliasi_material:
        type: integer
      qty:
        type: integer
    required:
    - id_rekonsiliasi_material
    - qty
    type: object
  model.IssueInventoryResponse:
    properties:
      balance:
        type: integer
      id_rekonsiliasi_material:
        type: integer
      previous_balance:
        type: integer
      qty_issued:
        type: integer
    type: object
  model.IssueInventorySuccessDoc:
    properties:
      data:
        $ref: '#/definitions/model.IssueInventoryResponse'
      message:
        example: inventory issued
        type: string
      status:
        example: success
        type: string
    type: object
...
  model.PackingListDetailResponse:
    properties:
      buyer:
        type: string
      created_at:
        type: string
      id_packing_list:
        type: integer
      id_surat_jalan_internal:
        type: integer
      id_wo:
        type: integer
      items:
        items:
          $ref: '#/definitions/model.PackingListItemResponse'
        type: array
      model:
        type: string
      total_garment_per_box:
        type: integer
      total_reject:
        type: integer
    type: object
  model.PackingListDetailSuccessDoc:
    properties:
      data:
        $ref: '#/definitions/model.PackingListDetailResponse'
      message:
        example: packing list retrieved
        type: string
      status:
        example: success
        type: string
    type: object
  model.PackingListItemResponse:
    properties:
      box_no_end:
        type: integer
      box_no_start:
        type: integer
      color:
        type: string
      created_at:
        type: string
      id_packing_list_item:
        type: integer
      note:
        type: string
      qty_box:
        type: integer
      qty_per_box:
        type: integer
      sizes:
        items:
          $ref: '#/definitions/model.PackingListItemSizeResponse'
        type: array
    type: object
  model.PackingListItemSizeResponse:
    properties:
      created_at:
        type: string
      id_packing_list_item_size:
        type: integer
      qty:
        type: integer
    type: object
  model.PackingListListItem:
    properties:
      buyer:
        type: string
      created_at:
        type: string
      id_packing_list:
        type: integer
      id_surat_jalan_internal:
        type: integer
      id_wo:
        type: integer
      model:
        type: string
      total_garment_per_box:
        type: integer
      total_reject:
        type: integer
    type: object
  model.PackingListListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/model.PackingListListItem'
        type: array
      pagination:
        $ref: '#/definitions/model.PaginationMeta'
    type: object
  model.PackingListListSuccessDoc:
    properties:
      data:
        $ref: '#/definitions/model.PackingListListResponse'
      message:
        example: packing lists retrieved
        type: string
      status:
        example: success
        type: string
    type: object
  model.PackingListResponse:
    properties:
      created_at:
        type: string
      id_packing_list:
        type: integer
      id_surat_jalan_internal:
        type: integer
      id_wo:
        type: integer
      items:
        items:
          $ref: '#/definitions/model.PackingListItemResponse'
        type: array
      total_garment_per_box:
        type: integer
      total_reject:
        type: integer
    type: object
  model.PackingListSuccessDoc:
    properties:
      data:
        $ref: '#/definitions/model.PackingListResponse'
      message:
        example: packing list created
        type: string
      status:
        example: success
        type: string
    type: object
  model.PaginationMeta:
    properties:
      limit:
        type: integer
      page:
        type: integer
      total_items:
        type: integer
      total_pages:
        type: integer
    type: object
...
  model.ReceiveInventoryRequest:
    properties:
      id_material_list:
        type: integer
      id_rekonsiliasi_material:
        type: integer
      keterangan:
        type: string
      qty:
        type: integer
      tanggal:
        type: string
    required:
    - id_material_list
    - id_rekonsiliasi_material
    - qty
    - tanggal
    type: object
  model.ReceiveInventoryResponse:
    properties:
      actual_kirim:
        type: integer
      balance:
        type: integer
      created_at:
        type: string
      id_material_list:
        type: integer
      id_received:
        type: integer
      id_rekonsiliasi_material:
        type: integer
      id_rekonsiliasi_material_terima:
        type: integer
      keterangan:
        type: string
      qty:
        type: integer
      tanggal:
        type: string
    type: object
  model.ReceiveInventorySuccessDoc:
    properties:
      data:
        $ref: '#/definitions/model.ReceiveInventoryResponse'
      message:
        example: inventory received
        type: string
      status:
        example: success
        type: string
    type: object
...
  model.SuratJalanClientDetailResponse:
    properties:
      created_at:
        type: string
      id_material_list:
        type: integer
      id_surat_jalan_client:
        type: integer
      id_wo:
        type: integer
      keterangan:
        type: string
      material_description:
        type: string
      qty:
        type: integer
      tanggal:
        type: string
    type: object
  model.SuratJalanClientDetailSuccessDoc:
    properties:
      data:
        $ref: '#/definitions/model.SuratJalanClientDetailResponse'
      message:
        example: surat jalan client retrieved
        type: string
      status:
        example: success
        type: string
    type: object
  model.SuratJalanClientListItem:
    properties:
      created_at:
        type: string
      id_material_list:
        type: integer
      id_surat_jalan_client:
        type: integer
      id_wo:
        type: integer
      keterangan:
        type: string
      material_description:
        type: string
      qty:
        type: integer
      tanggal:
        type: string
    type: object
  model.SuratJalanClientListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/model.SuratJalanClientListItem'
        type: array
      pagination:
        $ref: '#/definitions/model.PaginationMeta'
    type: object
  model.SuratJalanClientListSuccessDoc:
    properties:
      data:
        $ref: '#/definitions/model.SuratJalanClientListResponse'
      message:
        example: surat jalan clients retrieved
        type: string
      status:
        example: success
        type: string
    type: object
  model.SuratJalanInternalDetailResponse:
    properties:
      created_at:
        type: string
      id_surat_jalan_internal:
        type: integer
    type: object
  model.SuratJalanInternalDetailSuccessDoc:
    properties:
      data:
        $ref: '#/definitions/model.SuratJalanInternalDetailResponse'
      message:
        example: surat jalan internal retrieved
        type: string
      status:
        example: success
        type: string
    type: object
  model.SuratJalanInternalListItem:
    properties:
      created_at:
        type: string
      id_surat_jalan_internal:
        type: integer
    type: object
  model.SuratJalanInternalListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/model.SuratJalanInternalListItem'
        type: array
      pagination:
        $ref: '#/definitions/model.PaginationMeta'
    type: object
  model.SuratJalanInternalListSuccessDoc:
    properties:
      data:
        $ref: '#/definitions/model.SuratJalanInternalListResponse'
      message:
        example: surat jalan internals retrieved
        type: string
      status:
        example: success
        type: string
    type: object
  model.SuratJalanResponse:
    properties:
      created_at:
        type: string
      id_material_list:
        type: integer
      id_surat_jalan:
        type: integer
      keterangan:
        type: string
      qty:
        type: integer
      tanggal:
        type: string
      type:
        type: string
    type: object
  model.SuratJalanSuccessDoc:
    properties:
      data:
        $ref: '#/definitions/model.SuratJalanResponse'
      message:
        example: surat jalan created
        type: string
      status:
        example: success
        type: string
    type: object
...
  model.WarehouseErrorDetail:
    properties:
      code:
        example: related_data_not_found
        type: string
    type: object
  model.WarehouseErrorDoc:
    properties:
      error:
        $ref: '#/definitions/model.WarehouseErrorDetail'
      message:
        example: related data not found
        type: string
      status:
        example: error
        type: string
    type: object
  model.WarehouseValidationErrorDoc:
    properties:
      error:
        items:
          $ref: '#/definitions/response.ValidationErrorItem'
        type: array
      message:
        example: bad request
        type: string
      status:
        example: error
        type: string
    type: object
...
  response.ValidationErrorItem:
    properties:
      field:
        type: string
      message:
        type: string
      param:
        type: string
      rule:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@permatatex.local
    name: API Support
  description: API documentation for Permatatex IT Inventory System backend.
  license:
    name: MIT
  termsOfService: https://swagger.io/terms/
  title: Permatatex IT Inventory API
  version: "1.0"
paths:
...
  /api/v1/inventory/issue:
    post:
      consumes:
      - application/json
      description: Decrease reconciliation material balance when raw material is taken
        out from warehouse for production.
      parameters:
      - description: Issue inventory payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/model.IssueInventoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.IssueInventorySuccessDoc'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.WarehouseValidationErrorDoc'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.WarehouseErrorDoc'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/model.WarehouseErrorDoc'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.WarehouseErrorDoc'
      security:
      - BearerAuth: []
      summary: Issue Inventory
      tags:
      - Warehouse & Delivery
  /api/v1/inventory/receive:
    post:
      consumes:
      - application/json
      description: Record received inventory, insert reconciliation receipt detail,
        and update reconciliation balance in one database call.
      parameters:
      - description: Receive inventory payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/model.ReceiveInventoryRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.ReceiveInventorySuccessDoc'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.WarehouseValidationErrorDoc'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.WarehouseErrorDoc'
      security:
      - BearerAuth: []
      summary: Receive Inventory
      tags:
      - Warehouse & Delivery
...
  /api/v1/packing-lists:
    get:
      description: Returns a paginated list of packing list headers.
      parameters:
      - description: Page (default 1)
        in: query
        name: page
        type: integer
      - description: Limit (default 20)
        in: query
        name: limit
        type: integer
      - description: Search by buyer or model
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.PackingListListSuccessDoc'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.WarehouseErrorDoc'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.WarehouseErrorDoc'
      security:
      - BearerAuth: []
      summary: List Packing Lists
      tags:
      - Warehouse & Delivery
    post:
      consumes:
      - application/json
      description: Create packing list header with nested box items and size breakdowns
        in a single transaction.
      parameters:
      - description: Packing list payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/model.CreatePackingListRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.PackingListSuccessDoc'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.WarehouseValidationErrorDoc'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.WarehouseErrorDoc'
      security:
      - BearerAuth: []
      summary: Create Packing List
      tags:
      - Warehouse & Delivery
  /api/v1/packing-lists/{id}:
    get:
      description: Returns a single packing list with nested items and sizes.
      parameters:
      - description: Packing List ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.PackingListDetailSuccessDoc'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.WarehouseErrorDoc'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.WarehouseErrorDoc'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.WarehouseErrorDoc'
      security:
      - BearerAuth: []
      summary: Get Packing List Detail
      tags:
      - Warehouse & Delivery
...
  /api/v1/surat-jalan-clients:
    get:
      description: Returns a paginated list of client delivery notes.
      parameters:
      - description: Page (default 1)
        in: query
        name: page
        type: integer
      - description: Limit (default 20)
        in: query
        name: limit
        type: integer
      - description: Search by description or material
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.SuratJalanClientListSuccessDoc'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.WarehouseErrorDoc'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.WarehouseErrorDoc'
      security:
      - BearerAuth: []
      summary: List Surat Jalan Clients
      tags:
      - Warehouse & Delivery
  /api/v1/surat-jalan-clients/{id}:
    get:
      description: Returns a single client delivery note.
      parameters:
      - description: Surat Jalan Client ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.SuratJalanClientDetailSuccessDoc'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.WarehouseErrorDoc'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.WarehouseErrorDoc'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.WarehouseErrorDoc'
      security:
      - BearerAuth: []
      summary: Get Surat Jalan Client Detail
      tags:
      - Warehouse & Delivery
  /api/v1/surat-jalan-internals:
    get:
      description: Returns a paginated list of internal delivery notes.
      parameters:
      - description: Page (default 1)
        in: query
        name: page
        type: integer
      - description: Limit (default 20)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.SuratJalanInternalListSuccessDoc'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.WarehouseErrorDoc'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.WarehouseErrorDoc'
      security:
      - BearerAuth: []
      summary: List Surat Jalan Internals
      tags:
      - Warehouse & Delivery
  /api/v1/surat-jalan-internals/{id}:
    get:
      description: Returns a single internal delivery note.
      parameters:
      - description: Surat Jalan Internal ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.SuratJalanInternalDetailSuccessDoc'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.WarehouseErrorDoc'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.WarehouseErrorDoc'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.WarehouseErrorDoc'
      security:
      - BearerAuth: []
      summary: Get Surat Jalan Internal Detail
      tags:
      - Warehouse & Delivery
  /api/v1/surat-jalan/{type}:
    post:
      consumes:
      - application/json
      description: 'Create surat jalan document. Supported type: internal, client.'
      parameters:
      - description: Surat jalan type
        in: path
        name: type
        required: true
        type: string
      - description: Surat jalan client payload
        in: body
        name: payload
        schema:
          $ref: '#/definitions/model.CreateSuratJalanClientRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.SuratJalanSuccessDoc'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.WarehouseValidationErrorDoc'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.WarehouseErrorDoc'
      security:
      - BearerAuth: []
      summary: Create Surat Jalan
      tags:
      - Warehouse & Delivery
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"